refactor(client-runtime): share cloud delivery across clients
